[K5P-58] [refactor] 디렉토리 수정
