+++ src/main/java/site/billingwise/api/serverapi/domain/payment/dto/response/GetPaymentCardDto.java
@@ -1,16 +1,17 @@
-package site.billingwise.api.serverapi.domain.payment.dto.reponse;
+package site.billingwise.api.serverapi.domain.payment.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
+import site.billingwise.api.serverapi.domain.payment.dto.response.GetPaymentDto;
 
 @Getter
 @SuperBuilder
 @NoArgsConstructor
 @AllArgsConstructor
-public class GetPaymentAccountDto extends GetPaymentDto {
+public class GetPaymentCardDto extends GetPaymentDto {
     private String number;
-    private String bank;
+    private String company;
     private String owner;
 }
